requirements.txt
fastapi[all]
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
alembic